+++ config.go
@@ -12,8 +12,6 @@ import (
 // TODO: Merge #184 or #177 degree symbol (BartWillems:master, fleaz:master)
 // TODO: Merge #169 % option for network use (jrswab:networkPercentage)
 // TODO: Merge #167 configuration file (jrswab:configFile111)
-// TODO: Merge #166 Nord color scheme (jrswab:nordColorScheme)
-// TODO: Merge #165 README updates (theverything:add-missing-option-to-readme)
 // TODO: Merge #157 FreeBSD fixes & Nvidia GPU support (kraust:master)
 // TODO: Merge #156 Added temperatures for NVidia GPUs (azak-azkaran:master)
 // TODO: Merge #152 (implements #148 & #149) (mattLLVW:feature/network_interface_list)